fix: split preserve-input tests to satisfy backend-size limit
Move PRESERVE_INPUT and compatibility output-token tests into
transformer_preserve_input_test.go so transformer_test.go stays under
the 1000-line backend-size gate.
